backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
requests
python-multipart
python-dotenv
google-genai
